refactor: Update import paths for header component
