+++ apps/nextjs/src/app/(routes)/score-card/[id]/header.tsx
@@ -2,8 +2,8 @@ import { readUserSession } from "@/libs/auth";
 import { createSupabaseServerClient } from "@/libs/supabase/server";
 import { cn } from "@/libs/tailwind";
 
-import BackButton from "./back-button";
-import DeleteAlert from "./delete-alert";
+import BackButton from "../new/[id]/_components/back-button";
+import DeleteAlert from "../new/[id]/_components/delete-alert";
 
 const Header = async ({ gameId }: { gameId: string }) => {
   const session = await readUserSession();
