+++ apps/nextjs/src/app/(routes)/score-card/new/[id]/page.tsx
@@ -1,7 +1,7 @@
 import React, { Suspense } from "react";
 import { Loader2 } from "lucide-react";
 
-import Header from "./_components/header";
+import Header from "../../[id]/header";
 import ScoreCard from "./_components/score-card";
 
 const Page = ({ params }: { params: { id: string } }) => {
